Finished display map function in render.py
