+++ render.py
@@ -1,4 +1,5 @@
 import math
+import areas
 
 def print_bar(val, max):
     size = 20
@@ -15,3 +16,11 @@ def print_bar(val, max):
     for i in range(size - num):
       print(" ", end="")
     print("]", end="")
+  
+def display_map():
+    list_of_areas = [i.upper() for i in areas.game_map.keys()]
+    count = 1
+    print("Where to?\n")
+    for area in list_of_areas:
+      print(str(count) + ". " + area + "\n")
+      count += 1
