+++ rpg.py
@@ -31,19 +31,4 @@ class Player:
         print("FT: ", end="")
         render.print_bar(self.fatigue, 100)
 
-        print()
-
-class Locations:
-    def __init__(self, name, description):
-        self.name = name
-        self.description = description
-
-    def __repr__(self):
-        return "[{name}]".format(name=self.name)
-    
-    def print_description(self):
-        print()
-        print(self.description)
-        print()
-
-    
+        print()
